Cinema Project full basic Function

+++ src/main/java/com/spring/cinemaproject/Services/PaypalService.java
@@ -22,7 +22,7 @@ public class PaypalService {
     @Autowired
     private APIContext apiContext;
     public Payment createPayment(
-            Double total,
+            Float total,
             String currency,
             PaypalPaymentMethod method,
             PaypalPaymentIntent intent,
